Move utility functions/types to src/utils
