+++ apps/super-app/src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css';
 import { useEffect } from 'react';
 import CssBaseline from '@mui/material/CssBaseline';
-import { consoleSmokeTest } from './services';
+import { consoleSmokeTest } from './utils';
 import { ThemeProvider } from './common/contexts';
 import { AppRoutes } from './routes/router';
