+++ apps/super-app/src/services/index.ts
@@ -1,5 +1,3 @@
-export { consoleSmokeTest } from './CmsSmokeTest';
-
 export { getSupportDetailStrings } from './SupportDetail';
 export type {
   SupportDetailStrings,
@@ -12,8 +10,3 @@ export type { ServiceCardDetail } from './ServiceCardDetails';
 
 export { getTwc2Detail } from './ServiceSupportDetails';
 export type { Twc2SupportDetail } from './ServiceSupportDetails';
-
-export type { LanguageOption } from './Language';
-export { useLanguage } from './Language';
-export { getAppConfig } from './Config';
-export type { AppConfig } from './Config';
